+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## Added
 
-## Fixed
-
-## Changed
+- First prerelease version, implements `:funnel/whoami`, `:funnel/subscribe`,
+  `:funnel/unsubscribe`, `:funnel/broadcast`, `:funnel/query`.
+- Selectors: `true`, vector, map.